[RPA-606] Searching processes by tag (#405)
* feat(runbotics-orchestrator-ui): :sparkles: added searching by tag, creator and process name

* feat(runbotics-orchestrator-ui): :sparkles: added name, tag, creator highlighting during search

* feat(runbotics-orchestrator): :sparkles: deleting unsuned tags after process deletion

* chore: :pushpin: bump app version to 2.8.0-SNAPSHOT.77

* build(runbotics-orchestrator-ui): :green_heart: fixed type and file EOS

* refactor(runbotics-orchestrator-ui): :recycle: replaced tag filter function with reduce

* refactor(runbotics-orchestrator): :recycle: rewrite log in process find function

* refactor(runbotics-orchestrator-ui): :recycle: deleted unnecessary hook

* chore: :pushpin: bump app version to 2.8.0-SNAPSHOT.80

---------

+++ runbotics-orchestrator/src/main/java/com/runbotics/service/impl/ProcessServiceImpl.java
@@ -266,7 +266,15 @@ public class ProcessServiceImpl implements ProcessService {
     @Transactional
     public void delete(Long id) {
         log.debug("Request to delete Process : {}", id);
+        Optional<Process> process = processRepository.findById(id);
+        if (process.isEmpty()) {
+            throw new BadRequestAlertException("Cannot find process with this id", ENTITY_NAME, "processNotFound");
+        }
+
+        List<Long> remainingTags = process.get().getTags()
+            .stream().map(Tag::getId).collect(Collectors.toList());
         processRepository.deleteById(id);
+        tagService.deleteUnusedTags(remainingTags);
     }
 
     @Override
